+++ src/services/Redirects.php
@@ -11,7 +11,6 @@ namespace venveo\redirect\services;
 use Craft;
 use craft\base\Element;
 use craft\errors\ElementNotFoundException;
-use craft\events\DeleteElementEvent;
 use craft\events\ElementEvent;
 use craft\helpers\Db;
 use craft\helpers\ElementHelper;
@@ -23,7 +22,6 @@ use Throwable;
 use venveo\redirect\elements\db\RedirectQuery;
 use venveo\redirect\elements\Redirect;
 use venveo\redirect\Plugin;
-use venveo\redirect\queue\jobs\PruneDeletedRedirects;
 use venveo\redirect\records\Redirect as RedirectRecord;
 use yii\base\Component;
 use yii\base\ExitException;
@@ -133,7 +131,7 @@ class Redirects extends Component
         $destinationUrl = null;
 
         if ($redirect->type === Redirect::TYPE_STATIC) {
-            $processedUrl = $redirect->resolveDestinationUrl();
+            $processedUrl = $redirect->getDestinationUrl();
         } elseif ($redirect->type === Redirect::TYPE_DYNAMIC) {
             $sourceUrl = $redirect->sourceUrl;
             // Add leading and trailing slashes for RegEx
@@ -147,10 +145,14 @@ class Redirects extends Component
             if (preg_match('/\$[1-9]+/', $redirect->resolveDestinationUrl())) {
                 $processedUrl = preg_replace($sourceUrl, $redirect->resolveDestinationUrl(), $uri);
             } else {
-                $processedUrl = $redirect->resolveDestinationUrl();
+                $processedUrl = $redirect->getDestinationUrl();
             }
         } else {
             return;
+        }
+        if (!$processedUrl) {
+            Craft::warning('A matched redirect is missing a destination URL: '. $redirect->id);
+            throw new NotFoundHttpException();
         }
 
         // Saving elements takes a while - we're going to do our incrementing
@@ -222,13 +224,14 @@ class Redirects extends Component
             return;
         }
 
-        $redirect = new Redirect();
-        $redirect->siteId = $siteId;
-        $redirect->sourceUrl = $oldUri;
-        $redirect->destinationElementId = $elementId;
-        $redirect->destinationSiteId = $siteId;
-        $redirect->type = Redirect::TYPE_STATIC;
-        $redirect->statusCode = '301';
+            $redirect = new Redirect();
+            $redirect->siteId = $siteId;
+            $redirect->sourceUrl = $oldUri;
+            $redirect->destinationUrl = $newUri;
+            $redirect->destinationElementId = $elementId;
+            $redirect->destinationSiteId = $siteId;
+            $redirect->type = Redirect::TYPE_STATIC;
+            $redirect->statusCode = '301';
 
         try {
             Craft::$app->elements->saveElement($redirect);
@@ -244,58 +247,6 @@ class Redirects extends Component
         }
     }
 
-    /**
-     * When an element is deleted, we'll delete its redirects as well
-     *
-     * @param DeleteElementEvent $e
-     */
-    public function handleElementDeleted(DeleteElementEvent $e)
-    {
-        /** @var Element $element */
-        $element = $e->element;
-        if ($element instanceof Redirect) {
-            return;
-        }
-
-        if (!$element->uri) {
-            return;
-        }
-
-        $job = new PruneDeletedRedirects([
-            'deletedElementId' => $element->id,
-            'deletedElementUri' => $element->uri,
-            'siteId' => $element->siteId,
-            'hardDelete' => $e->hardDelete,
-        ]);
-        Craft::$app->queue->push($job);
-    }
-
-
-    /**
-     * When an element is restored, we'll try to restore any related redirects
-     *
-     * @param ElementEvent $e
-     * @throws Throwable
-     * @throws \yii\base\Exception
-     */
-    public function handleElementRestored(ElementEvent $e)
-    {
-        /** @var Element $element */
-        $element = $e->element;
-        if ($element instanceof Redirect) {
-            return;
-        }
-
-        if (!$element->uri) {
-            return;
-        }
-
-        // We'll search by URL and by ID just to be thorough
-        $redirectsByUrl = Redirect::find()->destinationUrl($element->uri)->siteId($element->siteId)->trashed(true)->all();
-        $redirectsById = Redirect::find()->destinationElementId($element->id)->siteId($element->siteId)->trashed(true)->all();
-        Craft::$app->elements->restoreElements($redirectsById);
-        Craft::$app->elements->restoreElements($redirectsByUrl);
-    }
 
     /**
      * Returns all sites that the user can create redirects in