+++ CHANGELOG.md
@@ -9,9 +9,13 @@
 
 ### Changed
 - Redirect source and destination URLs now get normalized when set on the element rather than on-save.
+- Removed element deletion redirect pruning
 
 ### Fixed
 - Fixed error caused by permissions & use new element permissions interface
+- Fixed ignored 404s getting pruned if not hit for a while
+- When a redirect is created for a changed slug, the original URI is still stored in addition to the element ID.
+- Redirects that are matched that don't have a valid destination now 404 properly.
 
 ## 4.0.0-beta.1 - 2021-3-19